feat(wallet): add BigDecimal ledger and secured mutations

+++ wallet-service/src/main/java/com/paypal/wallet_service/dto/HoldResponse.java
@@ -1,17 +1,19 @@
 package com.paypal.wallet_service.dto;
 
+import java.math.BigDecimal;
+
 public class HoldResponse {
     private String holdReference;
-    private Long amount;
+    private BigDecimal amount;
     private String status;
 
-    public HoldResponse(String holdReference, Long amount, String status) {
+    public HoldResponse(String holdReference, BigDecimal amount, String status) {
         this.holdReference = holdReference;
         this.amount = amount;
         this.status = status;
     }
 
     public String getHoldReference() { return holdReference; }
-    public Long getAmount() { return amount; }
+    public BigDecimal getAmount() { return amount; }
     public String getStatus() { return status; }
 }
